Now it mostly works
