+++ core/js/src/main/scala/com/monovore/decline/PlatformArguments.scala
@@ -1,0 +1,5 @@
+package com.monovore.decline
+
+abstract class PlatformArguments {
+
+}
